apps/agent-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adk-logistics-agent-server"
version = "0.1.0"
description = "Offline logistics agent workbench powered by Google ADK"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.135.1",
  "google-adk==1.27.2",
  "httpx==0.28.1",
  "pydantic==2.12.5",
  "python-dotenv==1.2.2",
  "uvicorn==0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "pytest-asyncio==1.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

